run_test/requirements.txt
acvl-utils==0.2
asttokens==2.0.5
backcall==0.2.0
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
batchgenerators==0.25
black==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.0.4
cmake==3.25.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
connected-components-3d==3.12.3
contourpy==1.1.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
dicom2nifti==2.4.8
dynamic-network-architectures==0.2
einops==0.6.1
entrypoints==0.4
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1692026125334/work
executing==0.8.2
filelock==3.9.0
fonttools==4.42.1
future==0.18.3
glob2==0.7
graphviz==0.20.1
idna==3.4
imagecodecs==2023.9.4
imageio==2.31.3
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
importlib-resources==6.0.1
ipykernel==6.9.1
ipython==8.0.1
itk==5.3.0
itk-core==5.3.0
itk-filtering==5.3.0
itk-io==5.3.0
itk-numerics==5.3.0
itk-registration==5.3.0
itk-segmentation==5.3.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.4.5
lazy_loader==0.3
linecache2==1.0.0
lit==15.0.7
MarkupSafe==2.1.2
matplotlib==3.8.0
matplotlib-inline==0.1.3
MedPy==0.4.0
monai==1.2.0
mpmath==1.2.1
mypy-extensions==0.4.3
nest-asyncio==1.5.4
networkx==3.0
nibabel==5.1.0
-e git+https://github.com/MIC-DKFZ/nnUNet.git@b4e97fe38a9eb6728077678d4850c41570a1cb02#egg=nnunetv2
numpy==1.24.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.1.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.1
prompt-toolkit==3.0.28
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1681775019467/work
ptyprocess==0.7.0
pure-eval==0.2.2
pydicom==2.4.3
Pygments==2.11.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-gdcm==3.0.22
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==22.3.0
requests==2.28.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
SimpleITK==2.3.0
six==1.16.0
smriprep-docker==0.9.2
stack-data==0.2.0
sympy==1.11.1
threadpoolctl==3.2.0
tifffile==2023.8.30
tomli==2.0.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.1
tqdm==4.66.1
traceback2==1.4.0
traitlets==5.1.1
triton==2.0.0
typing_extensions==4.1.1
tzdata==2023.3
unittest2==1.1.0
urllib3==1.26.13
vtk==9.2.6
wcwidth==0.2.5
yacs==0.1.8
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1689374466814/work
